Fix broken client date JSON serialization in RecordJsoImp l
Review at http://gwt-code-reviews.appspot.com/775801


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8559 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/requestfactory/client/impl/RecordJsoImpl.java
@@ -284,7 +284,7 @@ public class RecordJsoImpl extends JavaScriptObject implements Record {
     }
 
     if (value instanceof Date) {
-      double millis = ((Date) value).getTime();
+      long millis = ((Date) value).getTime();
       setString(property.getName(), String.valueOf(millis));
       return;
     }
